refactor(ic-asset): make error types public and add documentation
