+++ src/canisters/frontend/ic-asset/src/error/compute_evidence.rs
@@ -6,20 +6,26 @@ use crate::error::hash_content::HashContentError;
 use ic_agent::AgentError;
 use thiserror::Error;
 
+/// Errors related to computing evidence for a proposed update.
 #[derive(Error, Debug)]
 pub enum ComputeEvidenceError {
+    /// Failed when inspecting assets to be updated.
     #[error(transparent)]
     ProcessProjectAsset(#[from] CreateProjectAssetError),
 
+    /// Failed when determining which assets and encodings changed.
     #[error(transparent)]
     GatherAssetDescriptors(#[from] GatherAssetDescriptorsError),
 
+    /// Failed when reading assets properties from the asset canister.
     #[error(transparent)]
     GetAssetProperties(#[from] GetAssetPropertiesError),
 
+    /// Failed when computing hashes of asset content.
     #[error(transparent)]
     HashContent(#[from] HashContentError),
 
+    /// Failed to list assets in the asset canister.
     #[error("Failed to list assets: {0}")]
     ListAssets(AgentError),
 }
