+++ src/canisters/frontend/ic-asset/src/error/mod.rs
@@ -1,16 +1,35 @@
-pub mod compatibility;
-pub mod compute_evidence;
-pub mod create_chunk;
-pub mod create_encoding;
-pub mod create_project_asset;
-pub mod downgrade_commit_batch_arguments;
-pub mod gather_asset_descriptors;
-pub mod get_asset_config;
-pub mod get_asset_properties;
-pub mod hash_content;
-pub mod load_config;
-pub mod load_rule;
-pub mod prepare_sync_for_proposal;
-pub mod sync;
-pub mod upload;
-pub mod upload_content;
+//! Error types
+
+mod compatibility;
+mod compute_evidence;
+mod create_chunk;
+mod create_encoding;
+mod create_project_asset;
+mod downgrade_commit_batch_arguments;
+mod gather_asset_descriptors;
+mod get_asset_config;
+mod get_asset_properties;
+mod hash_content;
+mod load_config;
+mod load_rule;
+mod prepare_sync_for_proposal;
+mod sync;
+mod upload;
+mod upload_content;
+
+pub use compatibility::CompatibilityError;
+pub use compute_evidence::ComputeEvidenceError;
+pub use create_chunk::CreateChunkError;
+pub use create_encoding::CreateEncodingError;
+pub use create_project_asset::CreateProjectAssetError;
+pub use downgrade_commit_batch_arguments::DowngradeCommitBatchArgumentsV1ToV0Error;
+pub use gather_asset_descriptors::GatherAssetDescriptorsError;
+pub use get_asset_config::GetAssetConfigError;
+pub use get_asset_properties::GetAssetPropertiesError;
+pub use hash_content::HashContentError;
+pub use load_config::AssetLoadConfigError;
+pub use load_rule::LoadRuleError;
+pub use prepare_sync_for_proposal::PrepareSyncForProposalError;
+pub use sync::SyncError;
+pub use upload::UploadError;
+pub use upload_content::UploadContentError;
