+++ src/canisters/frontend/ic-asset/src/lib.rs
@@ -34,7 +34,7 @@
 mod asset;
 mod batch_upload;
 mod canister_api;
-mod error;
+pub mod error;
 mod evidence;
 mod sync;
 mod upload;
