+++ src/canisters/frontend/ic-asset/src/upload.rs
@@ -11,9 +11,9 @@ use crate::canister_api::methods::{
     list::list_assets,
 };
 use crate::canister_api::types::batch_upload::v0;
-use crate::error::compatibility::CompatibilityError::DowngradeV1TOV0Failed;
-use crate::error::upload::UploadError;
-use crate::error::upload::UploadError::{CommitBatchFailed, CreateBatchFailed, ListAssetsFailed};
+use crate::error::CompatibilityError::DowngradeV1TOV0Failed;
+use crate::error::UploadError;
+use crate::error::UploadError::{CommitBatchFailed, CreateBatchFailed, ListAssetsFailed};
 
 use ic_utils::Canister;
 use slog::{info, Logger};
